% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_strings.R
\name{concat_strings}
\alias{concat_strings}
\title{Concatenate Two Strings}
\usage{
concat_strings(x, y)
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector.}
}
\value{
A character vector with \code{x} and \code{y} pasted together elementwise.
}
\description{
Concatenates two character vectors elementwise using \code{paste()}.
}
\examples{
concat_strings("Hello", "World")
concat_strings(c("Good", "Bad"), c("Morning", "Luck"))
}
