% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_result_methods.R
\name{plot.concat_result}
\alias{plot.concat_result}
\title{Plot method for concat_result objects}
\usage{
\method{plot}{concat_result}(x, ...)
}
\description{
Plot method for concat_result objects
}
