% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_result_methods.R
\name{print.concat_result}
\alias{print.concat_result}
\title{Print method for concat_result objects}
\usage{
\method{print}{concat_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{concat_result}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for concat_result objects
}
