% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_result_methods.R
\name{summary.concat_result}
\alias{summary.concat_result}
\title{Summary method for concat_result objects}
\usage{
\method{summary}{concat_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{concat_result}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for concat_result objects
}
